Lo he hecho un poco mas responsive
ya no se van a tomar por culo los botones, he arreglado una mierda k
hacia k no se pudieran buscar productos por precio pork eran int y ahora
en la bd son float y le he puesto un hover al boton y aumentado el
tamaño

+++ Poxmania/src/main/java/es/sidelab/Poxmania/DataBaseUsage.java
@@ -41,7 +41,7 @@ public class DataBaseUsage implements CommandLineRunner {
 	@RequestMapping("/buscarprecio")
 	public ModelAndView porprecio(@RequestParam String min,@RequestParam String max) {
 		if(!min.isEmpty()&&!max.isEmpty()){
-			productos = repository.findByPrecioBetween(Integer.parseInt(min), Integer.parseInt(max));										
+			productos = repository.findByPrecioBetween(Float.parseFloat(min), Float.parseFloat(max));										
 			return new ModelAndView("index").addObject("productos", productos);
 		}else{
 			productos = repository.findAll();
